08-04-2025.1

+++ JavaQuest/JavaQuest1.java
@@ -57,9 +57,9 @@ public class JavaQuest1 {
     System.out.println("The variable varC is " + varC);
 
     // Reading Example
-    float fnumber = -130.20f; // By default, numbers are treated as double in Java.
+    double fnumber = -130.20d; // By default, numbers are treated as double in Java.
     // Use suffix 'f' with value to make it float.
-    double dnumber = -20.50;
+    float dnumber = -20.50f;
     System.out.println("The variable fnumber is " + fnumber);
     System.out.println("The variable dnumber is " + dnumber);
